Create README.md
Linear Regression, Churn Modelling, and Explanatory Data Analysis
This repository contains examples of linear regression, churn modelling, and explanatory data analysis algorithms implemented in Python.

Linear Regression
In the linear_regression directory, you will find an implementation of linear regression that uses the ordinary least squares (OLS) method to fit a linear model to a given dataset. The implementation includes a Jupyter notebook that demonstrates how the algorithm can be used to predict the price of a house based on various features, such as its size and location.

Churn Modelling
In the churn_modelling directory, you will find an implementation of a machine learning algorithm that can predict whether a customer will leave or stay with a given company. The implementation includes a Jupyter notebook that demonstrates how the algorithm can be trained on a dataset of customer attributes and transaction history to predict customer churn.

Explanatory Data Analysis
In the eda directory, you will find examples of how to perform exploratory data analysis (EDA) on a given dataset. The examples include visualizations, summary statistics, and data cleaning techniques that can be used to gain insights into a dataset and prepare it for further analysis.

Requirements
The code in this repository was written using Python 3.7. To run the examples, you will need to install the following packages:

NumPy
Pandas
Matplotlib
Seaborn
Scikit-learn
You can install these packages using pip:

Copy code
pip install numpy pandas matplotlib seaborn scikit-learn
Usage
To run the examples, simply open the Jupyter notebook in the appropriate directory and follow the instructions.

Contributing
If you find a bug or would like to contribute to this project, feel free to contact me domnicsparta92@gmail.com